refactoring code, from 'regulationLocation' to 'location'
