+++ templates/map.html.twig
@@ -18,7 +18,7 @@
             data-zoom="5"
 	    data-regulations-as-geojson-output-id="regulations_as_geojson_output"
 	    data-regulations-as-geojson-turbo-frame-id="regulations_as_geojson_turbo_frame"
-	    data-regulation-location-path="{{ path('get_regulation_location') }}"
+	    data-location-path="{{ path('get_location') }}"
             aria-hidden="true"
             />
 {% endblock %}
