+++ templates/regulation/fragments/_location.html.twig
@@ -1,5 +1,5 @@
 {% if location and measure %}
-<turbo-frame id="regulation_location_turbo_frame_{{ location.uuid }}" loading="lazy">
+<turbo-frame id="location_turbo_frame_{{ location.uuid }}" loading="lazy">
   <div class="app-card__header">
     <span class="app-card__img fr-icon-x-restriction fr-x-icon--primary fr-x-icon--sm" aria-hidden="true"></span>
     <span class="app-card__title fr-mb-0">
